+++ src/main/java/org/fdryt/ornamental/repository/FamilyRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FamilyRepository extends JpaRepository<Family, Long> {}
+public interface FamilyRepository extends JpaRepository<Family, Integer> {}
