+++ src/main/java/org/fdryt/ornamental/repository/PlantRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlantRepository extends JpaRepository<Plant, Long> {
+public interface PlantRepository extends JpaRepository<Plant, Integer> {
 
     @Query("""
             SELECT new org.fdryt.ornamental.dto.ProductResponseDTO
